+++ src/presentation/utils/email-validator-adapter.spec.ts
@@ -1,5 +1,5 @@
 import validator from 'validator';
-import { EmailValidatorAdapter } from './email-validator';
+import { EmailValidatorAdapter } from './email-validator-adapter';
 
 jest.mock('validator', () => ({
     isEmail (): boolean {
